// src/Moryx.FactoryMonitor.Web/app/src/app/api/models/cell-state-changed-model.ts
/*
 * Copyright (c) 2026 Phoenix Contact GmbH & Co. KG
 * Licensed under the Apache License, Version 2.0
*/

/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { CellState } from '../models/cell-state';
export interface CellStateChangedModel {
  id?: number;
  state?: CellState;
}