// src/Moryx.FactoryMonitor.Web/app/src/app/api/models/order-changed-model.ts
/*
 * Copyright (c) 2026 Phoenix Contact GmbH & Co. KG
 * Licensed under the Apache License, Version 2.0
*/

/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { InternalOperationClassification } from '../models/internal-operation-classification';
export interface OrderChangedModel {
  operation?: string | null;
  order?: string | null;
  state?: InternalOperationClassification;
}